requirements.txt
tensorflow>=2.8.0
streamlit>=1.22.0
streamlit-drawable-canvas>=0.9.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=9.0.0
requests>=2.26.0
python-dotenv>=0.19.0